+++ gsa/src/web/pages/tasks/__tests__/containerdialog.js
@@ -21,11 +21,7 @@ import React from 'react';
 import {render, fireEvent} from 'web/utils/testing';
 
 import ContainerDialog from '../containerdialog';
-import Task, {
-  AUTO_DELETE_KEEP,
-  AUTO_DELETE_KEEP_DEFAULT_VALUE,
-  AUTO_DELETE_NO,
-} from 'gmp/models/task';
+import Task from 'gmp/models/task';
 
 describe('ContainerDialog tests', () => {
   test('should render create dialog', () => {
@@ -75,8 +71,6 @@ describe('ContainerDialog tests', () => {
 
     expect(handleSave).toHaveBeenCalledWith({
       comment: 'lorem',
-      auto_delete: AUTO_DELETE_KEEP,
-      auto_delete_data: AUTO_DELETE_KEEP_DEFAULT_VALUE,
       in_assets: 1,
       id: undefined,
       name: 'ipsum',
@@ -104,12 +98,6 @@ describe('ContainerDialog tests', () => {
     const commentInput = getByName('comment');
     fireEvent.change(commentInput, {target: {value: 'lorem'}});
 
-    const autoDeleteKeepInput = getByName('auto_delete_data');
-    fireEvent.change(autoDeleteKeepInput, {target: {value: '10'}});
-
-    const [autoDeleteNoRadio] = queryAllByName('auto_delete');
-    fireEvent.click(autoDeleteNoRadio);
-
     const [, inAssetsNoRadio] = queryAllByName('in_assets');
     fireEvent.click(inAssetsNoRadio);
 
@@ -118,8 +106,6 @@ describe('ContainerDialog tests', () => {
 
     expect(handleSave).toHaveBeenCalledWith({
       comment: 'lorem',
-      auto_delete: AUTO_DELETE_NO,
-      auto_delete_data: 10,
       in_assets: 0,
       id: 't1',
       name: 'ipsum',